Changes in setManualBand data type for bands

+++ examples/test_conn/test_conn.ino
@@ -22,7 +22,7 @@ void setup()
     }
     delay(1000);
     Serial.println("=====Set bands to 8 and 20=====");
-    int bands[2] = {8, 20};
+    uint8_t bands[2] = {8, 20};
     if(quectel.setManualBand(2,bands))
     {
         Serial.println("Done!");
